Move add to type level
